+++ infrastructure/lib/src/shared_preferences_counter_repository.dart
@@ -1,0 +1,20 @@
+import 'package:domain/domain.dart';
+import 'package:shared_preferences/shared_preferences.dart';
+
+class SharedPreferencesCounterRepository implements CounterRepository {
+  static const _counterKey = "counterValue";
+
+  @override
+  Future<Counter?> get() async {
+    final prefs = await SharedPreferences.getInstance();
+    final retrievedValue = prefs.getInt(_counterKey);
+
+    return retrievedValue == null ? null : Counter(retrievedValue);
+  }
+
+  @override
+  Future save(Counter counter) async {
+    final prefs = await SharedPreferences.getInstance();
+    await prefs.setInt(_counterKey, counter.value);
+  }
+}
